chore: add create mirror
